fix on seeders for deploy

+++ routes/AdminUnidad_Medida/unidad_medida.php
@@ -9,19 +9,19 @@ use Illuminate\Support\Facades\Route;
 Route::controller(AdminUnidad_MedidaController::class)->group(function () {
 
     // GET METHOD
-    Route::get('/unidades', 'index')->middleware(['auth', 'verified', 'checkRole:Admin'])->name('unidades');
-    Route::get('/unidades/create', 'create')->middleware(['auth', 'verified', 'checkRole:Admin'])->name('unidad.create');
-    Route::get('/unidades/edit/{id}', 'edit')->middleware(['auth', 'verified', 'checkRole:Admin'])->name('unidad.edit');
+    Route::get('/unidades', 'index')->middleware(['auth', 'verified', 'checkRole:MegaAdmin,Admin'])->name('unidades');
+    Route::get('/unidades/create', 'create')->middleware(['auth', 'verified', 'checkRole:MegaAdmin,Admin'])->name('unidad.create');
+    Route::get('/unidades/edit/{id}', 'edit')->middleware(['auth', 'verified', 'checkRole:MegaAdmin,Admin'])->name('unidad.edit');
 
     // POST METHOD
-    Route::post('/unidades/create', 'store')->middleware(['auth', 'verified', 'checkRole:Admin'])->name('unidad.store');
+    Route::post('/unidades/create', 'store')->middleware(['auth', 'verified', 'checkRole:MegaAdmin,Admin'])->name('unidad.store');
 
 
     // PUT METHOD
-    Route::put('/unidades/update/{id}', 'update')->middleware(['auth', 'verified', 'checkRole:Admin'])->name('unidad.update');
-    Route::put('/unidades/unblock/{id}', 'unblock')->middleware(['auth', 'verified', 'checkRole:Admin'])->name('unidad.unblock');
+    Route::put('/unidades/update/{id}', 'update')->middleware(['auth', 'verified', 'checkRole:MegaAdmin,Admin'])->name('unidad.update');
+    Route::put('/unidades/unblock/{id}', 'unblock')->middleware(['auth', 'verified', 'checkRole:MegaAdmin,Admin'])->name('unidad.unblock');
 
 
     // DELETE METHOD
-    Route::delete('/unidades/destroy/{id}', 'destroy')->middleware(['auth', 'verified', 'checkRole:Admin'])->name('unidad.destroy');
+    Route::delete('/unidades/destroy/{id}', 'destroy')->middleware(['auth', 'verified', 'checkRole:MegaAdmin,Admin'])->name('unidad.destroy');
 });
